src/requirements.txt
Flask==2.0.1
openai==0.27.0
python-dotenv==0.19.0
PyPDF2==1.26.0
